services/invoice-service/pyproject.toml
[tool.poetry]
name = "selfmonitor-invoice-service"
version = "1.0.0"
description = "Enterprise invoicing and billing service for SelfMonitor FinTech platform"
authors = ["SelfMonitor Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.5.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"
weasyprint = "^60.2"
jinja2 = "^3.1.2"
httpx = "^0.25.2"
python-dateutil = "^2.8.2"
redis = "^5.0.1"
aiofiles = "^23.2.1"
pillow = "^10.1.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}
openpyxl = "^3.1.2"
pandas = "^2.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
factory-boy = "^3.3.0"
faker = "^20.1.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.mypy_cache
    | \.tox
    | \.venv
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = [
    "tests",
]
asyncio_mode = "auto"